Remove dead field & using; culture-invariant save timestamp; rename boot pipeline to BootPipeline and update uses; simplify BootScreen coroutine start
