+++ Assets/Scripts/Boot/BootPipeline.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e8704230fbde4d679968fcd27257329a
+guid: c0cdd587d50646f0b8ef9a3ca691570e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
