+++ Assets/Scripts/Core/Defs/Validation/DefValidator.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.IO;
 using System.Xml;
 
 namespace FantasyColony.Core.Defs.Validation {
